LPS-111943 Rename context/context to context/connection
