+++ modules/dxp/apps/analytics/analytics-reports-web/src/main/resources/META-INF/resources/js/context/connection.js
